refactor: refactor product list
