+++ src/pages/catalog-page/catalog/catalog.module.css
@@ -3,24 +3,10 @@
   width: 100%;
 }
 
-.cardsContainer {
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-  align-items: flex-start;
-  gap: 49px;
-  padding: 0 0 0 180px;
-}
-
 @media (width <=600px) {
   .container {
     flex-direction: column;
     align-items: center;
     gap: 40px;
-  }
-
-  .cardsContainer {
-    padding: 0;
-    justify-content: center;
   }
 }
