+++ src/pages/catalog-page/catalog/catalog.tsx
@@ -6,9 +6,7 @@ export const Catalog = () => {
   return (
     <div className={styles.container}>
       <SideBar />
-      <ul className={styles.cardsContainer}>
-        <ProductList />
-      </ul>
+      <ProductList />
     </div>
   );
 };
