+++ src/pages/catalog-page/product-list/product-list.module.css
@@ -1,55 +1,15 @@
-.cardContainer {
-  text-decoration: none;
-  color: inherit;
+.cardsContainer {
   display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  box-shadow: var(--shadow-2);
-  transition:
-    transform 0.3s ease,
-    box-shadow 0.3s ease;
-  cursor: pointer;
-  width: 255px;
-  gap: 21px;
-  padding: 15px;
-}
-
-.cardContainer:hover {
-  transform: scale(1.1);
-}
-
-li {
-  list-style: none;
-}
-
-.name {
-  font-size: 24px;
-  font-weight: 700;
-}
-
-.image {
+  flex-wrap: wrap;
   width: 100%;
-  height: 220px;
+  align-items: flex-start;
+  gap: 49px;
+  padding: 0 0 0 180px;
 }
 
-.description {
-  font-size: 20px;
-  font-weight: 400;
-}
-
-.priceContainer {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.price {
-  text-decoration: line-through;
-  font-size: 20px;
-}
-
-.discountPrice {
-  color: var(--clr-accent-1);
-  font-size: 20px;
+@media (width <=600px) {
+  .cardsContainer {
+    padding: 0;
+    justify-content: center;
+  }
 }
