+++ src/pages/catalog-page/product-list/product-list.tsx
@@ -1,15 +1,9 @@
 import styles from './product-list.module.css';
 import { catalogStore } from '../../../store/catalog-store.ts';
 import { observer } from 'mobx-react-lite';
-import { generatePath, Link } from 'react-router-dom';
-import { PagePath } from '../../../router/enums.ts';
 import { useEffect } from 'react';
-import DEFAULT_IMAGE from '../../../../assets/images/placeholder.png';
-import { CURRENCY_USD } from '../../../sources/constants/catalog.ts';
-import { DEFAULT_VALUE } from '../../../sources/enums/default-values.ts';
 import { messages } from '../../../sources/messages.ts';
-import { Button } from '../../../components/button/button.tsx';
-import { cartStore } from '../../../store/cart-store.ts';
+import { ProductCard } from './product-card/product-card.tsx';
 
 export const ProductList = observer(() => {
   const {
@@ -46,70 +40,14 @@ export const ProductList = observer(() => {
   }, []);
 
   return (
-    <>
+    <ul className={styles.cardsContainer}>
       {hasProducts ? (
-        productList.map(product => {
-          const isInCart = cartStore.isInCart(product.id);
-
-          const handleAddToCart = (
-            event: React.MouseEvent<HTMLButtonElement>
-          ) => {
-            event.preventDefault();
-            event.stopPropagation();
-            void cartStore.addItem({ productId: product.id, quantity: 1 });
-          };
-
-          return (
-            <li key={product.id}>
-              <Link
-                to={generatePath(PagePath.productPage, {
-                  categorySlug: product.categorySlug || DEFAULT_VALUE.CATEGORY,
-                  subcategorySlug:
-                    product.subcategorySlug || DEFAULT_VALUE.SUBCATEGORY,
-                  id: product.id,
-                })}
-                className={styles.cardContainer}
-              >
-                <img
-                  className={styles.image}
-                  src={product.image}
-                  alt={product.name}
-                  onError={event => {
-                    event.currentTarget.src = DEFAULT_IMAGE;
-                  }}
-                />
-                <p className={styles.name}>{product.name}</p>
-
-                <p className={styles.description}>{product.description}</p>
-                <div className={styles.priceContainer}>
-                  <p className={styles.price}>
-                    {CURRENCY_USD}
-                    {product.price}
-                  </p>
-                  <p className={styles.discountPrice}>
-                    {CURRENCY_USD}
-                    {product.discountPrice}
-                  </p>
-                </div>
-                {!isInCart ? (
-                  <Button onClick={handleAddToCart}>
-                    {messages.buttons.addToCart}
-                  </Button>
-                ) : (
-                  <>
-                    <p className={styles.inCartMessage}>
-                      {messages.productInCart}
-                    </p>
-                    <Button>{messages.buttons.removeFromCart}</Button>
-                  </>
-                )}
-              </Link>
-            </li>
-          );
-        })
+        productList.map(product => (
+          <ProductCard key={product.id} product={product} />
+        ))
       ) : (
         <p className={styles.emptyMessage}>{messages.noProducts}</p>
       )}
-    </>
+    </ul>
   );
 });
